+++ src/view/event/resizeEvent.ts
@@ -1,16 +1,11 @@
 /* eslint-disable max-lines-per-function */
 /* eslint-disable import/prefer-default-export */
 
+import { closeSearchForm } from '../getHeader';
+
 const changeSearchBoxHandler = (isMobile: boolean) => {
   if (window.innerWidth > 769 && isMobile) {
-    const button = document.querySelector('.mobile-search-button') as HTMLButtonElement;
-    const searchBox = document.querySelector('.search-box') as HTMLFormElement;
-
-    button.classList.add('mobile-visibility-visible');
-    button.classList.remove('mobile-visibility-hidden');
-
-    searchBox.classList.add('mobile-visibility-hidden');
-    searchBox.classList.remove('mobile-visibility-visible');
+    closeSearchForm();
   }
 };
